requirements.txt
gunicorn==23.0.0
whitenoise==6.8.2
requests==2.32.3
Django ==5.1.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
Pillow==11.0.0
django-environ== 0.11.2
django-filter ==  24.3
ruff == 0.8.0
black ==  24.10.0
pre-commit == 4.0.1
django-debug-toolbar == 4.4.6